% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{A dataset containing measurements on fish caught on the Tennessee River}
\format{
A data frame with 144 rows and 6 variables:
\describe{
\item{DDT}{ppm DDT in the flesh of fish}
\item{LENGTH}{Length of fish in cm}
\item{MILE}{Miles up river where fish is caught}
\item{RIVER}{River name}
\item{SPECIES}{Fish species}
\item{WEIGHT}{Weight of fish}
 }
}
\source{
{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
ddt
}
\description{
DDT data set
}
\details{
{This is a standard data set from the course MATH 4753 taken from the data sets provided by the text book.}
}
\examples{
data(ddt)
}
\keyword{datasets}
