% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybionomial.R
\name{mybin}
\alias{mybin}
\title{my binomial experiment}
\usage{
mybin(iter, n, p)
}
\arguments{
\item{iter}{variable containing iteration times}

\item{n}{variable indicating sample size}

\item{p}{variable indicating possibility}
}
\value{
plot
}
\description{
my binomial experiment
}
\examples{
\dontrun{mybin(10000,10,.7)}
}
