% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Title}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iteration times}

\item{x}{sample}

\item{fun}{function}

\item{alpha}{significance level}

\item{cx}{numeric character expansion facto}

\item{...}{other parameters}
}
\value{
histogram
}
\description{
Title
}
\examples{
\dontrun{myboot2(x=sam,fun="var",alpha=0.2)}


}
