% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{my clt poisson function}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{iteration times}

\item{lambda}{parameter for lambda}

\item{...}{other parameters}
}
\value{
histogram
}
\description{
my clt poisson function
}
\examples{
\dontrun{mycltp(n=2,iter=10000,lambda=4)}
}
