% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{my ddt plot function}
\usage{
myddt(df, SPECIES)
}
\arguments{
\item{df}{variable containing data set}

\item{SPECIES}{containing value of variable}
}
\value{
plot and table
}
\description{
my ddt plot function
}
\examples{
\dontrun{myddt(df = ddt, SPECIES = "CCATFISH")}
}
