% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhypergeometric.R
\name{myhyper}
\alias{myhyper}
\title{my hypergeometric simulation}
\usage{
myhyper(iter, N, r, n)
}
\arguments{
\item{iter}{variable containing iteration times}

\item{N}{variable indicating population size}

\item{r}{variable indicating number of success in population}

\item{n}{variable indicating sample size}
}
\value{
plot
}
\description{
my hypergeometric simulation
}
\examples{
\dontrun{myhyper(iter=100,N=20,n=5,r=12)}
}
