% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{mymlnorm function}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{vector of the sample}

\item{mu}{mean}

\item{sig}{standard deviation}

\item{...}{others}
}
\value{
graph of mean and standard deviation
}
\description{
mymlnorm function
}
\examples{
\dontrun{mymlnorm(x=c(),mu=seq(10,15,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)}
}
