% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{my distribution curve}
\usage{
myncurve(x, mu, sigma)
}
\arguments{
\item{x}{variable containing x value}

\item{mu}{variable indicating mean}

\item{sigma}{variable indicating sd}
}
\value{
plot
}
\description{
my distribution curve
}
\examples{
\dontrun{mycurve(mu=10,sigma=5,x=6)}

}
