% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{my sample}
\usage{
mysample(n, iter, time)
}
\arguments{
\item{n}{variable indicating sample size}

\item{iter}{variable containing iteration times}

\item{time}{run in the background}
}
\value{
plot
}
\description{
my sample
}
\examples{
\dontrun{mysample(n=1000,iter=30,time=1)}

}
