% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myscatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{my scatter histogram}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{variable containing x values}

\item{y}{variable containing y values}

\item{xlab}{label}

\item{ylab}{label}
}
\value{
plot
}
\description{
my scatter histogram
}
\examples{
\dontrun{with(ddt, scatterhist(LENGTH,WEIGHT, xlab="LENGTH"))}
}
