% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{my ddt plot function}
\usage{
myddt(df, cond)
}
\arguments{
\item{df}{variable containing dataset}

\item{cond}{variable containing condition to filter the dataset}
}
\value{
plot and table
}
\description{
my ddt plot function
}
\examples{
\dontrun{myddt(df = ddt, SPECIES = "CCATFISH")}
}
